Update dropdown menu display and styling
The dropdown menu in the header of the website has been updated to display its contents in a flex layout. Additional style enhancements have also been introduced to improve the visibility and interactions of the individual menu items. This includes setting properties such as font size, weight, color, and hover effects among others.

Signed-off-by: Arthur <arthur@thewandb.com>